Change the call to ghostscript to render pixmaps
Previously, ghostscript was called to produce a pcx file, which was then read in
with _read_pcx() and transformed to a pixmap. Now, -sDEVICE=bit... output of
ghostscript is used for the pixmap. The plan to do all in memory by
linking into the ghostscript library failed, since the callback functions
described in the ghostscript api, gsapi_set_stdio(), are not called in case an
output device is used. See https://www.ghostscript.com/doc/current/API.htm and
https://bugs.ghostscript.com/show_bug.cgi?id=694993.