Refactor code for reading image files (WIP)
Xfig would silently uncompress image files. Also, the open_picfile() and
close_picfile() functions in f_picobj.c are a half-abstracted layer to opening
regular files or streams from a pipe, passing back information to the user and
requesting the user to act based on this information. Replace these functions by
open_file(), close_file(), rewind_file() and uncompressed_file(). The latter
would create a temporary file from an compressed one, but does not modify the
original file. This is Work-In-Progress, since there is a muddle with opening
and closing in read_epsf_pdf().
Where source files are touched, the copyright comment on the top of the files is
aligned with the standard comment. Individual contributions are listed below the
standard comment, see, e.g., f_picobj.c, f_readgif.c or f_readjpg.c.