Correctly reset the locale when calling ghostscript
Before calling gs, either in gs_mediabox() or gs_exe(), which is called from
gs_bitmap(), all in u_ghostscript.c, the locale must be set to "C" or "POSIX".
This is for correctly parsing the decimal point of floating point numbers.
Afterwards, the locale is reset to the saved locale. Previously, the locale was
set to the environment values, which is not correct. The ghostscript functions
might be called when interactively creating a figure, but also when reading in a
fig file. However, during reading a fig file, the locale is and must also be set
to "C". Setting the locale to the environment, after a call to gs, would cause
reading of objects further down in the file to fail, because decimal points
could be incorrectly parsed.