Shift+wheel up/down pans left/right, ticket #95
This changes the user interface. The previous action of shift + mousewheel up,
popping up the edit panel, is now obtained by pressing the alt or meta key +
wheel up. Alt or meta + wheel down pops up the drawing panel.
The latter actions are obtained by Xt-translations, which are set in the
app-defaults file of xfig. Panning left or right is handled in the code.

*This commit changes the app-defaults file of xfig!* When using a version of
xfig after this commit, make sure the correct app-defaults file is loaded!